am able to edit ble characteristic values and hex, but can only read in the first byte

+++ ble5_simple_peripheral_cc2640r2lp_app/Application/simple_peripheral.c
@@ -1331,7 +1331,7 @@ static void SimplePeripheral_processGapMessage(gapEventHdr_t *pMsg)
  */
 static void SimplePeripheral_charValueChangeCB(uint8_t paramId)
 {
-  uint8_t *pValue = ICall_malloc(sizeof(uint8_t));
+  uint8_t *pValue = ICall_malloc(sizeof(uint16_t));
 
   if (pValue)
   {
